+++ src/components/sidebar/ConnectionCtrl.tsx
@@ -16,6 +16,8 @@ import { useConnectionCard } from '@/store/settings'
 import {
   ArrowDownCircleIcon,
   ArrowUpCircleIcon,
+  LinkIcon,
+  LinkSlashIcon,
   PauseIcon,
   PlayIcon,
   QuestionMarkCircleIcon,
@@ -53,7 +55,7 @@ export default defineComponent({
     const { t } = useI18n()
     const router = useRouter()
     const settingsModel = ref(false)
-    const { showTip } = useTooltip()
+    const { showTip, updateTip } = useTooltip()
     const { isLargeCtrlsBar } = useCtrlsBar(useConnectionCard.value ? 860 : 720)
 
     return () => {
@@ -166,6 +168,24 @@ export default defineComponent({
           <button
             class="btn btn-circle btn-sm"
             onClick={() => {
+              quickFilterEnabled.value = !quickFilterEnabled.value
+              updateTip(quickFilterEnabled.value ? t('showConnection') : t('hideConnection'))
+            }}
+            onMouseenter={(e) =>
+              showTip(e, quickFilterEnabled.value ? t('showConnection') : t('hideConnection'), {
+                appendTo: 'parent',
+              })
+            }
+          >
+            {quickFilterEnabled.value ? (
+              <LinkSlashIcon class="h-4 w-4" />
+            ) : (
+              <LinkIcon class="h-4 w-4" />
+            )}
+          </button>
+          <button
+            class="btn btn-circle btn-sm"
+            onClick={() => {
               isPaused.value = !isPaused.value
             }}
           >
@@ -214,7 +234,7 @@ export default defineComponent({
         </div>
       )
 
-      return <div class="glass-panel ctrls-bar">{content}</div>
+      return <div class="ctrls-bar">{content}</div>
     }
   },
 })